corrected errors due to byte alignment.

+++ src/common/inc/common.h
@@ -7,7 +7,7 @@
 #define FILE_PATH_LENGTH 192
 #define PK_PROC_NAME_LENGTH 32
 
-typedef uint8_t pk_inst_t;
+typedef uint16_t pk_inst_t;
 typedef char pk_name_t;
 
 typedef enum {
